Move get_python_lib and get_python_inc back to distutils.sysconfig
and deprecated them.
